+++ JsonRequest.Methods/Model/Order.cs
@@ -3,7 +3,7 @@
     public class Order
     {
         //var order = new Order { preIncidentId = 7, message = "Orden ingresada correctamente" };
-        public int preIncidentId { get; set; }
+        public long preIncidentId { get; set; }
         public string NroServicio { get; set; }
         public string message { get; set; }
     }
